moved epoch advancement logic into separate file
