+++ contracts/coconut-test/src/tests.rs
@@ -3,7 +3,7 @@
 
 mod deposit_and_release;
 mod helpers;
-mod redoing_upon_threshold;
+mod reset_upon_threshold;
 mod spend_credential_creates_proposal;
 mod submit_vk_creates_proposal;
 mod test_wrapper;
